[Api]: 'group' methods combine 'start' and 'selectRoles', schema remove 'additionalRoles'; Move utils to 'group' helpers. [Ui]: Split 'group_page' into panels.
